+++ src/main/java/com/example/stepperbackend/repository/MyExerciseRepository.java
@@ -1,8 +1,14 @@
 package com.example.stepperbackend.repository;
 
 import com.example.stepperbackend.domain.MyExercise;
+import org.springframework.data.domain.Page;
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
 
 public interface MyExerciseRepository extends JpaRepository<MyExercise, Long> {
 
+    Page<MyExercise> findByMember();
 }
+
