+++ src/main/java/com/example/stepperbackend/service/MyExerciseService.java
@@ -2,14 +2,19 @@ package com.example.stepperbackend.service;
 
 import com.example.stepperbackend.domain.MyExercise;
 import com.example.stepperbackend.repository.MyExerciseRepository;
-import com.example.stepperbackend.web.dto.MyExerciseRequestDTO;
+import com.example.stepperbackend.web.dto.MyExercise.MyExerciseRequestDTO;
+import jakarta.persistence.PersistenceContext;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+
+
+import java.util.List;
 
 @Service
 @RequiredArgsConstructor
 public class MyExerciseService {
     private final MyExerciseRepository myExerciseRepository;
+    @PersistenceContext
 
     public MyExercise addMyExercise(MyExerciseRequestDTO.AddExerciseDto request) {
         MyExercise myExercise = MyExercise.builder()
@@ -17,9 +22,14 @@ public class MyExerciseService {
                 .channel_name(request.getChannel_name())
                 .video_image(request.getVideo_image())
                 .video_title(request.getVideo_title())
-                .body_part(request.getBodyPart())
+                .body_part(request.getBody_part())
                 .build();
 
         return myExerciseRepository.save(myExercise);
+    }
+
+    public MyExercise checkMyExercise(MyExerciseRequestDTO.CheckExerciseDto request) {
+
+
     }
 }
