--- src/main/java/com/example/stepperbackend/web/dto/MyExerciseResponseDTO.java
@@ -1,17 +1,0 @@
-package com.example.stepperbackend.web.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-public class MyExerciseResponseDTO {
-    @Builder
-    @Getter
-    @NoArgsConstructor
-    @AllArgsConstructor
-    public static class AddExerciseDTO{
-        String status;
-        String message;
-    }
-}
